BoomBox Sound Playe Change

+++ src/main/java/net/morimori/imp/packet/ServerSoundStreamMessage.java
@@ -6,10 +6,10 @@ import net.morimori.imp.sound.WorldSoundKey;
 public class ServerSoundStreamMessage {
 	public String key;
 	public WorldSoundKey wsk;
-	public int offsetpos;
+	public long offsetpos;
 	public boolean stop;
 
-	public ServerSoundStreamMessage(String key, WorldSoundKey wsk, int offsetpos, boolean stop) {
+	public ServerSoundStreamMessage(String key, WorldSoundKey wsk, long offsetpos, boolean stop) {
 		this.key = key;
 		this.wsk = wsk;
 		this.offsetpos = offsetpos;
@@ -19,7 +19,7 @@ public class ServerSoundStreamMessage {
 	public static ServerSoundStreamMessage decodeMessege(PacketBuffer buffer) {
 		return new ServerSoundStreamMessage(buffer.readString(32767),
 				new WorldSoundKey(buffer.readString(32767), buffer.readString(32767), buffer.readString(32767)),
-				buffer.readInt(), buffer.readBoolean());
+				buffer.readLong(), buffer.readBoolean());
 	}
 
 	public static void encodeMessege(ServerSoundStreamMessage messegeIn, PacketBuffer buffer) {
@@ -27,7 +27,7 @@ public class ServerSoundStreamMessage {
 		buffer.writeString(messegeIn.wsk.getFolder());
 		buffer.writeString(messegeIn.wsk.getName());
 		buffer.writeString(messegeIn.wsk.getUUID());
-		buffer.writeInt(messegeIn.offsetpos);
+		buffer.writeLong(messegeIn.offsetpos);
 		buffer.writeBoolean(messegeIn.stop);
 
 	}
